Generative AI/company_chatbot/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
chromadb
gradio
pypdf
python-dotenv
